chore: Keep Settings and wizard presentation sources under 500 lines (#1787)
