+++ Packages/src/Editor/Presentation/Setup/SetupWizardSkillsWorkflowController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f941f9579e2c4ca0a3090b762b65261c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
