+++ Packages/src/Editor/Presentation/Setup/SetupWizardWorkflowController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a51f9558400452da951800bb00a4aa6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
