+++ Packages/src/Editor/Presentation/Setup/ThirdPartyToolMigrationWizardWorkflowController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f148ff5a135443f9b980b8e076988243
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
