+++ Packages/src/Editor/Presentation/UIToolkit/Components/ToolListRowData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a032ce2e589496b96f98d44d1acde05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
