+++ Packages/src/Editor/Presentation/UIToolkit/Components/ToolSettingsSectionListViewController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8218bd41b6a24ab9b053956a5e921d45
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
