+++ Packages/src/Editor/Presentation/UIToolkit/Components/ToolSettingsSectionRowBinder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6c2af33f8d1429ea6464639e0fe676e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
